Add (back) testing of table report format
